gitignore update, audio functionality test
