+++ .gitignore
@@ -2,3 +2,4 @@
 *.mp4
 ./Documents
 ./Test_Images
+./build
