+++ .vscode/launch.json
@@ -6,8 +6,7 @@
       "type": "cppdbg",
       "request": "launch",
       //"args": ["./Documents/RickRoll.mp4", "test_video", "test_audio"],
-      "args": ["./Documents/RickRoll.mp4"],
-      //"args": [],
+      "args": [],
       "stopAtEntry": false,
       "externalConsole": false,
       "cwd": "c:/Users/David/Documents/DevProjects/msys64/home/David/Workspace/VideoPlayer",
